refactor: Move CompressionUtils to polars-utils
