+++ crates/polars-parquet/src/parquet/parquet_bridge.rs
@@ -1,5 +1,6 @@
 // Bridges structs from thrift-generated code to rust enums.
 
+pub use polars_utils::compression::*;
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
@@ -134,125 +135,6 @@ impl From<CompressionOptions> for CompressionCodec {
             CompressionOptions::Zstd(_) => CompressionCodec::ZSTD,
             CompressionOptions::Lz4Raw => CompressionCodec::LZ4_RAW,
         }
-    }
-}
-
-/// Defines valid compression levels.
-pub(crate) trait CompressionLevel<T: std::fmt::Display + std::cmp::PartialOrd> {
-    const MINIMUM_LEVEL: T;
-    const MAXIMUM_LEVEL: T;
-
-    /// Tests if the provided compression level is valid.
-    fn is_valid_level(level: T) -> Result<(), ParquetError> {
-        let compression_range = Self::MINIMUM_LEVEL..=Self::MAXIMUM_LEVEL;
-        if compression_range.contains(&level) {
-            Ok(())
-        } else {
-            Err(ParquetError::InvalidParameter(format!(
-                "valid compression range {}..={} exceeded.",
-                compression_range.start(),
-                compression_range.end()
-            )))
-        }
-    }
-}
-
-/// Represents a valid brotli compression level.
-#[derive(Debug, Eq, PartialEq, Hash, Clone, Copy)]
-pub struct BrotliLevel(u32);
-
-impl Default for BrotliLevel {
-    fn default() -> Self {
-        Self(1)
-    }
-}
-
-impl CompressionLevel<u32> for BrotliLevel {
-    const MINIMUM_LEVEL: u32 = 0;
-    const MAXIMUM_LEVEL: u32 = 11;
-}
-
-impl BrotliLevel {
-    /// Attempts to create a brotli compression level.
-    ///
-    /// Compression levels must be valid.
-    pub fn try_new(level: u32) -> Result<Self, ParquetError> {
-        Self::is_valid_level(level).map(|_| Self(level))
-    }
-
-    /// Returns the compression level.
-    pub fn compression_level(&self) -> u32 {
-        self.0
-    }
-}
-
-/// Represents a valid gzip compression level.
-#[derive(Debug, Eq, PartialEq, Hash, Clone, Copy)]
-pub struct GzipLevel(u8);
-
-impl Default for GzipLevel {
-    fn default() -> Self {
-        // The default as of miniz_oxide 0.5.1 is 6 for compression level
-        // (miniz_oxide::deflate::CompressionLevel::DefaultLevel)
-        Self(6)
-    }
-}
-
-impl CompressionLevel<u8> for GzipLevel {
-    const MINIMUM_LEVEL: u8 = 0;
-    const MAXIMUM_LEVEL: u8 = 9;
-}
-
-impl GzipLevel {
-    /// Attempts to create a gzip compression level.
-    ///
-    /// Compression levels must be valid (i.e. be acceptable for [`flate2::Compression`]).
-    pub fn try_new(level: u8) -> Result<Self, ParquetError> {
-        Self::is_valid_level(level).map(|_| Self(level))
-    }
-
-    /// Returns the compression level.
-    pub fn compression_level(&self) -> u8 {
-        self.0
-    }
-}
-
-#[cfg(feature = "gzip")]
-impl From<GzipLevel> for flate2::Compression {
-    fn from(level: GzipLevel) -> Self {
-        Self::new(level.compression_level() as u32)
-    }
-}
-
-/// Represents a valid zstd compression level.
-#[derive(Debug, Eq, PartialEq, Hash, Clone, Copy)]
-pub struct ZstdLevel(i32);
-
-impl CompressionLevel<i32> for ZstdLevel {
-    // zstd binds to C, and hence zstd::compression_level_range() is not const as this calls the
-    // underlying C library.
-    const MINIMUM_LEVEL: i32 = 1;
-    const MAXIMUM_LEVEL: i32 = 22;
-}
-
-impl ZstdLevel {
-    /// Attempts to create a zstd compression level from a given compression level.
-    ///
-    /// Compression levels must be valid (i.e. be acceptable for [`zstd::compression_level_range`]).
-    pub fn try_new(level: i32) -> Result<Self, ParquetError> {
-        Self::is_valid_level(level).map(|_| Self(level))
-    }
-
-    /// Returns the compression level.
-    pub fn compression_level(&self) -> i32 {
-        self.0
-    }
-}
-
-#[cfg(feature = "zstd")]
-impl Default for ZstdLevel {
-    fn default() -> Self {
-        Self(zstd::DEFAULT_COMPRESSION_LEVEL)
     }
 }
